Move some methods into other modules
